requirements.txt
django==3.1.5
djangorestframework==3.13.1
django-elasticsearch-dsl-drf==0.22.1
psycopg2>=2.8
numpy
lxml==4.6.3
requests==2.25.1
Markdown==3.3.4
pandas==1.2.5
tqdm==4.61.1
opencv-contrib-python==4.5.5.64
matplotlib==3.3.4
pillow==8.2.0
pdf2image==1.16.0
pytesseract==0.3.9
imutils==0.5.4
djongo==1.3.6
pymongo==3.11.4
elasticsearch==6.8.2
elasticsearch-dsl==7.3.0
python-dotenv
